Factor lexer into something that can implement yylex easier
From the Ragel docs (Chapter 5):

> The Ragel code generator is very flexible. The generated code has no
> dependencies and can be inserted in any function, perhaps inside a
> loop if desired. The user is responsible for declaring and
> initializing a number of required variables, including the current
> state and the pointer to the input stream. These can live in any
> scope. Control of the input processing loop is also possible: the user
> may break out of the processing loop and return to it at any time.

We take advantage of this: the `return` that we've written in the Ragel
actions returns from where the `%% write exec;` was written. The problem
is that if we return early, it happens before the Ragel processing loop
would have advanced the `p` pointer (the next character to read), so we
have a helper function that makes sure to advance it ourselves after
returning.